requirements.txt
alembic==1.7.7
black==22.3.0
click==8.0.3
Flask==2.0.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
SQLAlchemy==1.4.36
tomli==2.0.1
typing-extensions==4.2.0
Werkzeug==2.0.2
zipp==3.8.0
